+++ lib/ecto/exceptions.ex
@@ -131,7 +131,14 @@ defmodule Ecto.CastError do
   @moduledoc """
   Raised when a changeset can't cast a value.
   """
-  defexception [:message]
+  defexception [:message, :type, :value]
+
+  def exception(opts) do
+    type  = Keyword.fetch!(opts, :type)
+    value = Keyword.fetch!(opts, :value)
+    msg   = opts[:message] || "cannot cast #{inspect value} to #{inspect type}"
+    %__MODULE__{message: msg, type: type, value: value}
+  end
 end
 
 defmodule Ecto.InvalidURLError do